coredump-gui: implement reporting to KDE
this turned out a bit invasive but is mostly straight forward.

instead of manually tracking dependencies between members the Patient
now has a FaultContext that gets initialized at the same time with the
entire bundle of variables. for simplicity that includes some KDE
specific stuff instead of complicating live with subclasses and casting
and what not.

a whole bunch of stuff in the cpp got renamed.

m_faultContext initialization now uses the new drkonqi-core library to
check if the crash was KDE software and then creates a suitable context.

report() invokes drkonqi similar to drkonqi-coredump-launcher.

error strings and return types have been adjusted accordingly.

BUG: 511524