cmake: new feature option WITH_DRKONQI_REPORTING
this controls whether the drkonqi based UX is enabled. we'd like
distributions that ship older software to disable drkonqi so the user
doesn't report bugs against unmaintained software versions.

since this would require huge amounts of infrastructure to pull off
automatically we instead offer a single knob to turn drkonqi reporting
off.

we now have global UX to guide the user to distribution bug trackers
even when they have no crash reporting tool of their own, allowing for
crashes to not go unnoticed regardless of the drkonqi main UI being
involved or not.

to be clear: this only disables the drkonqi based KDE bugzilla/sentry
UX. the general coredump tech remains active and picks up the slack.

notably this should be enabled for distributions that follow the
releases of KDE software relatively closely. distributions that do not
should either disable the reporting outright or disable it when the
software stack goes out of support

BUG: 501946