developerpage: compress text updates
previously we'd stream all lines into the textarea even when we were
about to replace them with a complete trace in onStateChanged. this
caused excessive laggyness when the trace was a) fast and b) verbose.

instead we now stream the lines into a temporary array and flush them
every 500ms to the textarea. that should improve UI responsiveness while
loading trace considerably.