drkonqidialog: untangle widgets pieces
we still have a widgets ui in case the quick ui fails to run for some
reason (not sure that can still happen but that's besides the point).
previously the widgets ui was always loaded in background which also
meant that it always was calculating UI updates when traces changed.
this caused excessive extra load for absolutely no reason.

instead move the entire widgets tech into its own class. only initialize
this class in the quick error conditions.