globalnotifiertruck: split notification text and differentiate exe from unit name
this resolves a problem that is similar to the terminalemulator problem
handled in the serviceForUnitName mapping.

e.g. firefox launches plasma-browser-integration's native messaging host
app so it is inside the firefox cgroup (assuming no flatpak). this would
then incorrectly report firefox crashed when it was indeed pbi that
crashed.

resolving this situation is a bit difficult becuase we don't really know
what is and isn't directly a part of the service. instead we now display
both the executable name and the service name when it is an app (pbi
(part of Firefox) has encountered a fatal error and was closed)